+++ production/deploy/gcp/terraform/environment/demo/buyer/buyer.tf
@@ -16,7 +16,6 @@ locals {
   gcp_project_id = "" # Example: "your-gcp-project-123"
   environment    = "" # Must be <= 3 characters. Example: "abc"
   image_repo     = "" # Example: "us-docker.pkg.dev/your-gcp-project-123/services"
-
 }
 
 provider "google" {
@@ -57,6 +56,7 @@ module "buyer" {
     ENABLE_BUYER_FRONTEND_BENCHMARKING = "" # Example: "false"
     CREATE_NEW_EVENT_ENGINE            = "" # Example: "false"
     ENABLE_BIDDING_COMPRESSION         = "" # Example: "true"
+    ENABLE_PROTECTED_APP_SIGNALS       = "" # Example: "false"
     # This flag should only be set if console.logs from the AdTech code(Ex:generateBid()) execution need to be exported as VLOG.
     # Note: turning on this flag will lead to higher memory consumption for AdTech code execution
     # and additional latency for parsing the logs.
@@ -70,12 +70,13 @@ module "buyer" {
     #    "enableBuyerDebugUrlGeneration": true,
     #    "enableAdtechCodeLogging": false,
     #  }"
-    JS_NUM_WORKERS      = "" # Example: "64" Must be <=vCPUs in bidding_machine_type.
-    JS_WORKER_QUEUE_LEN = "" # Example: "200".
-    JS_WORKER_MEM_MB    = "" # Example: "3072" JS_WORKER_MEM_MB/JS_WORKER_QUEUE_LEN > average JS request size.
-    ROMA_TIMEOUT_MS     = "" # Example: "10000"
-    TELEMETRY_CONFIG    = "" # Example: "mode: EXPERIMENT"
-    COLLECTOR_ENDPOINT  = "" # Example: "collector-buyer-1-${local.environment}.bfe-gcp.com:4317"
+    JS_NUM_WORKERS            = "" # Example: "64" Must be <=vCPUs in bidding_machine_type.
+    JS_WORKER_QUEUE_LEN       = "" # Example: "200".
+    ROMA_TIMEOUT_MS           = "" # Example: "10000"
+    TELEMETRY_CONFIG          = "" # Example: "mode: EXPERIMENT"
+    COLLECTOR_ENDPOINT        = "" # Example: "collector-buyer-1-${local.environment}.bfe-gcp.com:4317"
+    ENABLE_OTEL_BASED_LOGGING = "" # Example: "false"
+    CONSENTED_DEBUG_TOKEN     = "" # Example: "<unique_id>"
 
     # Reach out to the Privacy Sandbox B&A team to enroll with Coordinators and update the following flag values.
     # More information on enrollment can be found here: https://github.com/privacysandbox/fledge-docs/blob/main/bidding_auction_services_api.md#enroll-with-coordinators