+++ services/buyer_frontend_service/buyer_frontend_main.cc
@@ -132,12 +132,16 @@ absl::StatusOr<TrustedServersConfigClient> GetConfigClient(
   config_client.SetFlag(FLAGS_consented_debug_token, CONSENTED_DEBUG_TOKEN);
   config_client.SetFlag(FLAGS_enable_otel_based_logging,
                         ENABLE_OTEL_BASED_LOGGING);
+  config_client.SetFlag(FLAGS_enable_protected_app_signals,
+                        ENABLE_PROTECTED_APP_SIGNALS);
 
   if (absl::GetFlag(FLAGS_init_config_client)) {
     PS_RETURN_IF_ERROR(config_client.Init(config_param_prefix)).LogError()
         << "Config client failed to initialize.";
   }
 
+  VLOG(1) << "Protected App Signals support enabled on the service: "
+          << config_client.GetBooleanParameter(ENABLE_PROTECTED_APP_SIGNALS);
   VLOG(1) << "Successfully constructed the config client.\n";
   return config_client;
 }
@@ -184,18 +188,19 @@ absl::Status RunServer() {
   server_common::InitTelemetry(
       config_util.GetService(), kOpenTelemetryVersion.data(),
       telemetry_config.TraceAllowed(), telemetry_config.MetricAllowed(),
-      config_client.GetBooleanParameter(ENABLE_OTEL_BASED_LOGGING));
-  server_common::ConfigureMetrics(CreateSharedAttributes(&config_util),
-                                  CreateMetricsOptions(), collector_endpoint);
+      telemetry_config.LogsAllowed() &&
+          config_client.GetBooleanParameter(ENABLE_OTEL_BASED_LOGGING));
   server_common::ConfigureTracer(CreateSharedAttributes(&config_util),
                                  collector_endpoint);
   server_common::ConfigureLogger(CreateSharedAttributes(&config_util),
                                  collector_endpoint);
   AddSystemMetric(metric::BfeContextMap(
       std::move(telemetry_config),
-      opentelemetry::metrics::Provider::GetMeterProvider()
-          ->GetMeter(config_util.GetService(), kOpenTelemetryVersion.data())
-          .get()));
+      server_common::ConfigurePrivateMetrics(
+          CreateSharedAttributes(&config_util),
+          CreateMetricsOptions(telemetry_config.metric_export_interval_ms()),
+          collector_endpoint),
+      config_util.GetService(), kOpenTelemetryVersion.data()));
 
   BuyerFrontEndService buyer_frontend_service(
       std::make_unique<HttpBiddingSignalsAsyncProvider>(