+++ services/common/util/error_categories.h
@@ -54,6 +54,7 @@ inline constexpr char kMalformedBuyerInput[] = "Malformed buyer input.";
 // Server side errors listed here.
 inline constexpr char kInternalError[] = "Internal Error";
 inline constexpr char kEncryptionFailed[] = "Encryption Failure.";
+inline constexpr char kRequestCancelled[] = "Request Cancelled by Client.";
 
 // Error handling related constants.
 inline constexpr char kErrorDelimiter[] = "; ";